Pretty close on the logic
Turns out the data I spent all that time on last night sucks. The French
distribution doesn't even include Morocco or Algeria! Absolute garbage.
And the distribution was not what I thought. I thought it was the
distribution of speakers within the country, but it's actually the
distribution of world speakers within that country.

Therefore I'm switching to using the wiki chart. Gonna pause on this
logic and switch to that.